% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwarfplanets.R
\docType{data}
\name{dwarfplanets}
\alias{dwarfplanets}
\title{Specifications of some of the dwarf planets in the Solar System.}
\format{A data frame with 3 records and 3 variables:
\describe{
  \item{name}{ dwarf planet name;}
  \item{distance}{length of orbital semi-major axis [units: AU];}
  \item{eccentricity}{orbital eccentricity.}
  \item{mass}{dwarf planet mass [units: kg];}
  \item{period}{orbital period [units: earth years];}
}}
\source{
These data were constructed using information from:
https://en.wikipedia.org/wiki/Eris_(dwarf_planet)
https://en.wikipedia.org/wiki/Haumea
https://en.wikipedia.org/wiki/Makemake
http://www.space.com/18584-dwarf-planets-solar-system-infographic.html
}
\usage{
dwarfplanets
}
\description{
A dataset containing the name, orbital radius and orbital eccentricity of
the dwarf planets in the Solar System.
}
\examples{
plot(eccentricity ~ distance, data = dwarf.planets)
}
\keyword{datasets}
