% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{ellipse_area}
\alias{ellipse_area}
\title{Calculate the area of an ellipse.}
\usage{
ellipse_area(r, e = 0)
}
\arguments{
\item{r}{Length of the semi-major axis.}

\item{e}{The eccentricity.}
}
\value{
The area of the ellipse with specified semi-major axis and eccentricity.
}
\description{
Calculate the area of an ellipse.
}
\examples{
ellipse_area(1, 1)
ellipse_area(1)
}
