% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_foci.R
\name{ellipse_foci_distance}
\alias{ellipse_foci_distance}
\title{Calculate distance from the center to the focal points}
\usage{
ellipse_foci_distance(a, b)
}
\arguments{
\item{a}{length of the major axis}

\item{b}{length of the minor axis}
}
\value{
The the distance of the focal point to the center
}
\description{
Calculate distance from the center to the focal points
}
\examples{
ellipse_foci_distance(1, 0.5)
}
