% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi_minor.R
\name{ellipse_semi_minor}
\alias{ellipse_semi_minor}
\title{Calculate the semi-minor axis of an ellipse.}
\usage{
ellipse_semi_minor(a, b, f)
}
\arguments{
\item{a}{Distance from one foci to edge of ellipse.}

\item{b}{Distance from other foci to the same point on the edge of the ellipse.}

\item{f}{Distance between foci}
}
\value{
The length of the semi-minor axis.
}
\description{
Calculate the semi-minor axis of an ellipse.
}
\examples{
ellipse_semi_minor(1, 3, 6)
ellips_semi_minor(1, 2, 3)
}

