% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moons.R
\docType{data}
\name{moons}
\alias{moons}
\title{Specifications of the moons of each planet in the Solar System.}
\format{A data frame with 7 records and 4 variables: # TODO: ADD ALL 149 OBSERVABLE MOONS OF THE SOLAR SYSTEM
\describe{
  \item{moon}{moon name;}
  \item{planet}{name of the planet it orbits;}
  \item{distance}{length of orbital semi-major axis [units: AU];}
  \item{eccentricity}{orbital eccentricity.}
}}
\source{
These data were constructed using information from NASA's Planetary Fact Sheet at http://nssdc.gsfc.nasa.gov/planetary/factsheet/.
}
\usage{
moons
}
\description{
A dataset containing the name, oribiting planet name, orbital radius and orbital eccentricity of
the moons in the Solar System.
}
\examples{
\dontrun{
plot(eccentricity ~ distance, data = planets)
}
}
\keyword{datasets}
