% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planets.R
\docType{data}
\name{planets}
\alias{planets}
\title{Specifications of the planets in the Solar System.}
\format{A data frame with 9 records and 3 variables:
\describe{
  \item{name}{planet name;}
  \item{distance}{length of orbital semi-major axis [units: AU];}
  \item{eccentricity}{orbital eccentricity.}
  \item{mass}{planet mass [units: kg];}
  \item{period}{orbital period [units: earth years];}
}}
\source{
These data were constructed using information from NASA's Planetary Fact Sheet at http://nssdc.gsfc.nasa.gov/planetary/factsheet/.
}
\usage{
planets
}
\description{
A dataset containing the name, orbital radius and orbital eccentricity of
the planets in the Solar System. I'm old school, so Pluto is still in there.
Added in two more planets that are considered one of the more Earth-like alien planets. (see more at http://www.space.com/30172-six-most-earth-like-alien-planets.html)
}
\examples{
plot(eccentricity ~ distance, data = planets)
}
\keyword{datasets}
