% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precession.R
\name{precession_newt}
\alias{precession_newt}
\title{calculation of theoretical perihelion precession of celestial bodies using Newtonian mechanics}
\source{
The formula used in calculating precession was taken from Richard Fitzpatrick's physics notes. http://farside.ph.utexas.edu/teaching/336k/Newtonhtml/node115.html#e13.126
}
\usage{
precession_newt()
}
\description{
calculation of theoretical perihelion precession of celestial bodies using Newtonian mechanics
}
