% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramanujan.R
\name{ramanujan}
\alias{ramanujan}
\title{Calculates Perimeter using Ramanujan Approximation}
\usage{
ramanujan(a, b = a)
}
\arguments{
\item{a}{#semi-major axis}

\item{b}{#semi-minor axis}
}
\value{
#circumference of circle/ellipse
}
\description{
Calculates Perimeter using Ramanujan Approximation
}
\examples{
ramanujan(10,12) = 69.25791
ramanujan(6,7) = 40.90114
}
