% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moons.R
\docType{data}
\name{moons}
\alias{moons}
\title{Specifications of Jupiter's Galilean moons.}
\format{A data frame with 4 records and 3 variables:
\describe{
  \item{name}{moon name;}
  \item{distance}{length of orbital semi-major axis [units: AU];}
  \item{eccentricity}{orbital eccentricity.}
}}
\source{
These data were constructed using information from http://www.pa.msu.edu/people/horvatin/Astronomy_Facts/planet_pages/Jupiters_moons.htm
}
\usage{
moons
}
\description{
A dataset containing the name, orbital radius and orbital eccentricity of
Jupiter's Galilean moons.
}
\examples{
plot(eccentricity ~ distance, data = moons)
}
\keyword{datasets}
